backend/requirements.txt
fastapi
python-multipart
pillow
numpy
tensorflow
opencv-python
uvicorn